+++ src/main/java/com/example/solidconnection/auth/service/oauth/KakaoOAuthService.java
@@ -1,7 +1,8 @@
-package com.example.solidconnection.auth.service;
+package com.example.solidconnection.auth.service.oauth;
 
 import com.example.solidconnection.auth.client.KakaoOAuthClient;
 import com.example.solidconnection.auth.dto.oauth.OAuthUserInfoDto;
+import com.example.solidconnection.auth.service.SignInService;
 import com.example.solidconnection.siteuser.domain.AuthType;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
 import org.springframework.stereotype.Service;
